refactor: Revert Svd code and add a caching layer for V matrix.
For some bizarre reason the SVD decomposition code only works using the original
code. The Eigen3 implementation introduces some different relative sign on the
columns of U and V, I assume that this sign instability leads leads to problems
in the SVD stack algorithm. For now it's better to leave it as it is.

The computation of the matrix V requires a multiple over all the V matrix on the
stack. To prevent recomputing them all the time we introduce a V matrix cache.
This cache is invalidated only when we pop from the stack, since this would
require a matrix inversion and we don't want to do that. For pushing onto the
stack is fine because this only requires a matrix multiplication.