fix: Resize might alloc inside temporary pool
Due to new the way we handle measurement methods, temporary objects might have
different sizes, and when acquired will require an allocation.

This is unfortunate because the purpose of the temporary pool is to not
allocate. We should keep it since this simplify a lot of the code and does not
increase the memory usage significantly.